requirements.txt
appdirs==1.4.3
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.1
django-allauth==0.36.0
django-cors-headers==2.2.0
django-heroku==0.3.1
django-rest-auth==0.9.3
djangorestframework==3.9.3
filesystems==0.22.1
gunicorn==19.9.0
idna==2.8
mkenv==3.4.0
mysql-connector-python==8.0.16
mysqlclient==1.4.2.post1
oauthlib==3.0.1
packaging==19.0
Pillow==6.0.0
protobuf==3.8.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
pyparsing==2.4.0
pyrsistent==0.15.2
python-decouple==3.1
python3-openid==3.1.0
pytoml==0.1.20
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
tqdm==4.32.1
urllib3==1.25.2
virtualenv==16.6.0
virtualenvwrapper-win==1.2.5
whitenoise==4.1.2
zope.interface==4.6.0
